chore: move filter_events
